+++ sources/packages/backend/apps/api/src/route-controllers/student/models/student.dto.ts
@@ -91,7 +91,7 @@ export class UpdateStudentAPIInDTO
 }
 
 /**
- * Student AEST search parameters.
+ * Student search parameters.
  */
 export class StudentSearchAPIInDTO {
   @ValidateIf(
