--- sources/packages/web/src/constants/search-types.ts
@@ -1,4 +1,0 @@
-export enum SearchTypes {
-  Ministry = "ministry",
-  Institution = "institution",
-}
