+++ sources/packages/web/src/types/contracts/institution/InstitutionUser.ts
@@ -70,6 +70,7 @@ export interface InstitutionStateForStore {
   operatingName: string;
   institutionType: string;
   isBCPrivate: boolean;
+  isBCPublic: boolean;
   /**
    * Indicates if the institution has a BCeID business guid
    * associated with, if not it is a basic BCeID institution.
